Make networkInterfaces work with dual stack environments, add ipIgnoreList
This makes networkInterfaces dual stack compatible by removing the usage of nice_interfaces_get_ip_for_interface, which used ioctl and SIOCGIFADDR which are not AF_INET6 compatible

This was done via glibc`s getifaddrs. The pitfall here is that, differently from ioctl, this fetches link local and private IPs from interfaces, which brought in the need of manually filtering link local addresses and the introduction of the ipIgnoreList configuration

The ipIgnoreList config is a list of IP addresses (v4-6) to be ignored during the gathering phase. With that, you can prevent adding IPs from the desired interfaces you know are useless to libnice`s agent